fix: move tests into same structure as contracts
