+++ tests/core/citycoin-auth.test.ts
@@ -1,8 +1,8 @@
-import { assertEquals, describe, types, run, Chain, beforeEach, it} from "../deps.ts";
-import { AuthModel } from "../models/auth.model.ts";
-import { CoreModel } from "../models/core.model.ts";
-import { TokenModel } from "../models/token.model.ts";
-import { Accounts, Context } from "../src/context.ts";
+import { assertEquals, describe, types, run, Chain, beforeEach, it} from "../../deps.ts";
+import { AuthModel } from "../../models/auth.model.ts";
+import { CoreModel } from "../../models/core.model.ts";
+import { TokenModel } from "../../models/token.model.ts";
+import { Accounts, Context } from "../../src/context.ts";
 
 let ctx: Context;
 let chain: Chain;
