+++ tests/core/citycoin-core-v1.test.ts
@@ -1,7 +1,7 @@
-import { assertEquals, describe, TxReceipt, types, run, Chain, beforeEach, it } from "../deps.ts";
-import { CoreModel } from "../models/core.model.ts";
-import { TokenModel } from "../models/token.model.ts";
-import { Accounts, Context } from "../src/context.ts";
+import { assertEquals, describe, TxReceipt, types, run, Chain, beforeEach, it } from "../../deps.ts";
+import { CoreModel } from "../../models/core.model.ts";
+import { TokenModel } from "../../models/token.model.ts";
+import { Accounts, Context } from "../../src/context.ts";
 
 
 let ctx: Context;
