+++ tests/vote/citycoin-vote-v1.test.ts
@@ -1,8 +1,8 @@
-import { describe, assertEquals, types, Account, run, Chain, it, beforeEach} from "../deps.ts";
-import { CoreModel } from "../models/core.model.ts";
-import { TokenModel } from "../models/token.model.ts";
-import { VoteModel } from "../models/vote.model.ts";
-import { Accounts, Context } from "../src/context.ts";
+import { describe, assertEquals, types, Account, run, Chain, it, beforeEach} from "../../deps.ts";
+import { CoreModel } from "../../models/core.model.ts";
+import { TokenModel } from "../../models/token.model.ts";
+import { VoteModel } from "../../models/vote.model.ts";
+import { Accounts, Context } from "../../src/context.ts";
 
 let ctx: Context;
 let chain: Chain;
